Fix MCP Lambda error header import
